requirements.txt
facebook-business==17.0.4
notion-client==2.2.1
python-dotenv==1.0.1
requests==2.31.0
